Switches to Ansible roles
I have been learning more about ansible and how to make smaller modules
that can be covered with tests. TDD for provisioning!

This PR reorganizes everything into Roles and each Role is configured to
use molecule for testing. The goal is to be able to iterate on small
bits of functiinality (eg: go tools or aws cli install) and test them in
isolation while still having a top level playbook for installing
everything I want in my Linux laptop setup.