Fixes usage of ansible_facts. Removes references to macOS (#3)
I was using ansible_facts wrong. For example, when fetching the arch,
the correct way is to use `ansible_facts['architecture']` and I was
using `ansible_architecture` which is deprecated. And the calls to
`ansible_distrubition` were always returning empty. This is now fixed.

I also removed all references to macOS because 1) this is a linux
playbook and 2) the claims I was making were wrong. In GitHub actions,
the `ansible_facts['distribution']` was coming back as `MacOSX` and not
`Darwin` as I had thought. And that stuff is better documented in the
repo for the multi-os-dev proof of concept - no need to say anything
about it in the comments of this code.