Fixes broken CI by using specific versions of linters
A naive approach just to see if I can force the versions of molecule,
ansible, and the linters to be something specific. Just going with the
latest introduced breaking changes where the code didn't change in the
repo.

A better approach would be to use the same requirements.txt file for
installing the dependencies locally as what is used to install the
dependencies in the GitHub Action workflows. But, that seems more
involved so I am punting on that until I have time to mess with a global
requirements.txt file. I am just not sure how to tell GitHub workflows
on where to pickup the file.