Removes ruby install (#9)
The Ruby install ended up being a sudo install where all bundle commands
etc had to be executed with sudo and it just didn't work. I used that
molecule module instead of installing rvm because i didn't think
Rubymine would behanve well with rvm, but it ends up it has first class
support for rvm. So yeah, fucked it on that one.