Fixes bashrc issues
The dotfiles were not connecting correctly when the playbook was run
locally using --ask-become-pass. AND changing .bash_profile is actually
folly since it only affects the login shell which is not what is used
when I fire up a gnome-terminal on Rocky. Moved the file edits to
bashrc.

Tested in molecule and running ansible-playbook locally.