Sets global playbook become to an env variable and molecule default
Solves the issue where running the playbook locally requires that the
main playbook be run with become: true or dotfiles and the tpm and vim
settings end up editing the incorrect bashrc file (root instead of the
user who ran the playbook). But, if become: true while running in
molecule on Github Actions will result in the tests failing. I wanted
become: true to be the default for running the playbook for it's main
intention, which is to configure my dev machine, The desired outcome has
been created here where the molecule tests are configured one way by
setting the env variable in the provisioner in the
molecule/default/molecule.yml file.